Upgrade to Stylelint v17 and latest dependencies
 stylelint                                   16.12.0  →  17.6.0
 and
 @stylistic/stylelint-config                 2.0.0  →   5.0.0
 @stylistic/stylelint-plugin                 3.1.1  →   5.1.0
 postcss-html                                1.7.0  →   1.8.1
 stylelint-config-recommended               14.0.1  →  18.0.0
 stylelint-no-unsupported-browser-features   8.0.2  →   8.1.1
 and supporting
 eslint-config-wikimedia                     0.28.2 →   0.31.0"

In order to make the upgrade work, we're:
- Adding Node and ESM metadata to 'package.json':
-- adding "exports" to expose the package entry points; Stylelint's
   migration docs explicitly recommend adding it when moving to ESM.
-- adding "engines" to specify the minimum required by Stylelint v17.
- Upgrading 'no-at-import-css' to ESM and using the new API for it.
- Massively updating tests (that was mostly Claude's work here).
- Disabling a number of new rules in order to keep downstream upgrade
  work feasible.

Leaving 'index.js' in CommonJS reduces the chance of surprising
downstream users with package-level ESM changes.

Also follow a common 'package.json' order by
help of 'npx sort-package-json'.

Also unifying on official spelling of "Less"
(it's "Less", not more, not "LESS") in comments and messages.

Fixing some typos here and there.

Bug: #259
