+++ test/fixtures/default/valid.less
@@ -14,6 +14,13 @@ div {
 // Off: no-invalid-position-at-import-rule
 @import 'importAfterRule.less';
 
+// Off: string-no-newline
+// Off: comment-word-disallowed-list
+// Off: property-no-deprecated
+// Off: property-no-unknown
+// Off: media-feature-name-value-no-unknown
+// Off: declaration-property-value-no-unknown
+
 /* Check customSyntax is working */
 p {
 	div {
