+++ test/fixtures/modern/valid.less
@@ -14,6 +14,9 @@ div {
 // Off: no-invalid-position-at-import-rule
 @import 'importAfterRule.less';
 
+// Off: media-feature-name-value-no-unknown
+// Off: declaration-property-value-no-unknown
+
 /* Check customSyntax is working */
 p {
 	div {
